generate_deck/pyproject.toml
[project]
authors = [
  {name = "Salvatore D'angelo", email = "sasadangelo@gmail.com"},
]
description = "Utility scripts for flashcards app"
name = "generate-deck"
version = "0.1.0"

[tool.poetry.dependencies]
diffusers = "^0.34.0"
gtts = "^2.5.4"
openai = "^1.97.0"
pillow = "^11.3.0"
python = ">=3.9,<4.0"
python-dotenv = "^1.1.1"
requests = "^2.31.0"
torch = "^2.7.1"
transformers = "^4.53.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
